requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
faiss-cpu
python-dotenv
tiktoken
pytesseract
Pillow
edge-tts
faster-whisper
sounddevice
pygame
scipy
numpy
